Set sync start without rewind.
Move GetDifficultyRatio into txhelpers.
