+++ txhelpers/txhelpers.go
@@ -5,8 +5,11 @@ package txhelpers
 
 import (
 	"fmt"
+	"math/big"
 	"sort"
+	"strconv"
 
+	"github.com/decred/dcrd/blockchain"
 	"github.com/decred/dcrd/chaincfg"
 	"github.com/decred/dcrd/chaincfg/chainhash"
 	"github.com/decred/dcrd/txscript"
@@ -194,4 +197,24 @@ func MedianCoin(s []float64) float64 {
 		return s[middle]
 	}
 	return (s[middle] + s[middle-1]) / 2
+}
+
+// GetDifficultyRatio returns the proof-of-work difficulty as a multiple of the
+// minimum difficulty using the passed bits field from the header of a block.
+func GetDifficultyRatio(bits uint32, params *chaincfg.Params) float64 {
+	// The minimum difficulty is the max possible proof-of-work limit bits
+	// converted back to a number.  Note this is not the same as the proof of
+	// work limit directly because the block difficulty is encoded in a block
+	// with the compact form which loses precision.
+	max := blockchain.CompactToBig(params.PowLimitBits)
+	target := blockchain.CompactToBig(bits)
+
+	difficulty := new(big.Rat).SetFrac(max, target)
+	outString := difficulty.FloatString(8)
+	diff, err := strconv.ParseFloat(outString, 64)
+	if err != nil {
+		fmt.Printf("Cannot get difficulty: %v", err)
+		return 0
+	}
+	return diff
 }
